Relocate classes in app module
