+++ app/src/main/java/akio/apps/myrun/MyRunApp.kt
@@ -7,8 +7,8 @@ import akio.apps.myrun.feature.configurator.ConfiguratorFacade
 import akio.apps.myrun.feature.tracking.RouteTrackingService
 import akio.apps.myrun.log.CrashReportTree
 import akio.apps.myrun.log.MyDebugTree
-import akio.apps.myrun.wiring.AppComponent
-import akio.apps.myrun.wiring.DaggerAppComponent
+import akio.apps.myrun.di.AppComponent
+import akio.apps.myrun.di.DaggerAppComponent
 import akio.apps.myrun.worker.AppMigrationWorker
 import akio.apps.myrun.worker.UpdateUserRecentPlaceWorker
 import android.app.Application
