+++ app/src/main/java/akio/apps/myrun/feature/splash/SplashActivity.kt
@@ -4,10 +4,9 @@ import akio.apps.myrun.data.authentication.api.model.SignInSuccessResult
 import akio.apps.myrun.feature.base.DialogDelegate
 import akio.apps.myrun.feature.base.lifecycle.collectEventRepeatOnStarted
 import akio.apps.myrun.feature.base.viewmodel.lazyViewModelProvider
-import akio.apps.myrun.feature.home.HomeActivity
+import akio.apps.myrun.feature.main.MainActivity
 import akio.apps.myrun.feature.registration.SignInActivity
-import akio.apps.myrun.feature.splash.wiring.DaggerSplashFeatureComponent
-import android.content.Context
+import akio.apps.myrun.feature.splash.di.DaggerSplashFeatureComponent
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
@@ -67,7 +66,7 @@ class SplashActivity : AppCompatActivity() {
     }
 
     private fun goHome() {
-        startActivity(HomeActivity.clearTaskIntent(this))
+        startActivity(MainActivity.clearTaskIntent(this))
     }
 
     private fun verifySignInResult(resultCode: Int, data: Intent?) {
@@ -88,9 +87,5 @@ class SplashActivity : AppCompatActivity() {
         private const val RC_ON_BOARDING = 2
 
         private const val SPLASH_MIN_SHOWTIME = 700
-
-        fun clearTaskIntent(context: Context) = Intent(context, SplashActivity::class.java).apply {
-            addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK or Intent.FLAG_ACTIVITY_NEW_TASK)
-        }
     }
 }
